# feat(text-to-image): integrate NSFW safety checker