--- a/runner/openapi.json
+++ b/runner/openapi.json
@@ -269,7 +269,7 @@
                     "safety_check": {
                         "type": "boolean",
                         "title": "Safety Check",
-                        "default": false
+                        "default": true
                     },
                     "seed": {
                         "type": "integer",
@@ -443,7 +443,7 @@
                     "safety_check": {
                         "type": "boolean",
                         "title": "Safety Check",
-                        "default": false
+                        "default": true
                     },
                     "seed": {
                         "type": "integer",