--- a/runner/openapi.yaml
+++ b/runner/openapi.yaml
@@ -165,6 +165,10 @@ components:
           type: string
           title: Negative Prompt
           default: ''
+        safety_check:
+          type: boolean
+          title: Safety Check
+          default: false
         seed:
           type: integer
           title: Seed
@@ -258,10 +262,14 @@ components:
         seed:
           type: integer
           title: Seed
+        nsfw:
+          type: boolean
+          title: Nsfw
       type: object
       required:
       - url
       - seed
+      - nsfw
       title: Media
     TextToImageParams:
       properties:
@@ -286,6 +294,10 @@ components:
           type: string
           title: Negative Prompt
           default: ''
+        safety_check:
+          type: boolean
+          title: Safety Check
+          default: false
         seed:
           type: integer
           title: Seed