This pull request integrates the [CompVis/stable-diffusion-safety-checker](https://huggingface.co/CompVis/stable-diffusion-safety-checker) into the `text-to-image` and `image-to-image` pipelines. By enabling the `safety_check` input variable, users can detect the generation of NSFW images. The feature is enabled by default on Orchestrators, but they can opt to perform the safety check on the CPU using the `SAFETY_CHECKER_DEVICE` environment variable to conserve VRAM, though this may slow down operations.

I have decided against implementing this feature in the `image-to-video` pipeline as checking each frame would introduce significant latency, and I believe responsibility for checking the input image should lie with the dApp. However, I can readily add a check on the input image if desired.

### Performance Results

#### T2I pipeline

Benchmarking was conducted to assess how the new feature impacts inference speed and VRAM usage:

**Commands:**

```bash
docker run --gpus 0 -v /home/ricks/.lpData/models:/models livepeer/ai-runner:latest python bench.py --pipeline text-to-image --model_id stabilityai/sd-turbo --runs 10 --batch_size 3
```

and

```bash
docker run -e  SAFETY_CHECKER_DEVICE=<DEVICE> --gpus 0 -v /home/ricks/.lpData/models:/models livepeer/ai-runner:latest python bench.py --pipeline image-to-image --model_id stabilityai/sd-turbo --runs 10 --batch_size 3
```

**Original Metrics:**

```bash
pipeline load time: 1.664s
pipeline load max GPU memory allocated: 4.824GiB
pipeline load max GPU memory reserved: 4.904GiB
avg inference time: 0.635s
avg inference time per output: 0.212s
avg inference max GPU memory allocated: 7.651GiB
avg inference max GPU memory reserved: 9.521GiB
```

**With Safety Checker (GPU, safety_check not set):**

```bash
pipeline load time: 2.586s
pipeline load max GPU memory allocated: 5.966GiB
pipeline load max GPU memory reserved: 6.025GiB
avg inference time: 0.632s
avg inference time per output: 0.211s
avg inference max GPU memory allocated: 8.793GiB
avg inference max GPU memory reserved: 10.623GiB
```

**With Safety Checker (GPU, safety_check set):**

```bash
pipeline load time: 2.279s
pipeline load max GPU memory allocated: 5.966GiB
pipeline load max GPU memory reserved: 6.025GiB
avg inference time: 0.695s
avg inference time per output: 0.232s
avg inference max GPU memory allocated: 8.793GiB
avg inference max GPU memory reserved: 10.623GiB
```

**With Safety Checker (CPU, safety_check not set):**

```bash
pipeline load time: 2.245s
pipeline load max GPU memory allocated: 4.824GiB
pipeline load max GPU memory reserved: 4.904GiB
avg inference time: 0.640s
avg inference time per output: 0.213s
avg inference max GPU memory allocated: 7.651GiB
avg inference max GPU memory reserved: 9.521GiB
```

**With Safety Checker (CPU, safety_check set):**

```bash
pipeline load time: 2.203s
pipeline load max GPU memory allocated: 4.832GiB
pipeline load max GPU memory reserved: 4.895GiB
avg inference time: 1.233s
avg inference time per output: 0.411s
avg inference max GPU memory allocated: 7.659GiB
avg inference max GPU memory reserved: 9.512GiB
```

#### I2I pipeline

Benchmarking was conducted to assess how the new feature impacts inference speed and VRAM usage:

**Commands:**

```bash
docker run --gpus 0 -v /home/ricks/.lpData/models:/models livepeer/ai-runner:latest python bench.py --pipeline image-to-image --model_id stabilityai/sd-turbo --runs 10 --batch_size 3
```

and

```bash
docker run -e  SAFETY_CHECKER_DEVICE=<DEVICE> --gpus 0 -v /home/ricks/.lpData/models:/models livepeer/ai-runner:latest python bench.py --pipeline image-to-image --model_id stabilityai/sd-turbo --runs 10 --batch_size 3
```

**Original Metrics:**

```bash
pipeline load time: 1.672s
pipeline load max GPU memory allocated: 4.825GiB
pipeline load max GPU memory reserved: 4.900GiB
avg inference time: 4.179s
avg inference time per output: 1.393s
avg inference max GPU memory allocated: 16.092GiB
avg inference max GPU memory reserved: 20.023GiB
```

**With Safety Checker (GPU, safety_check not set):**

```bash
pipeline load time: 2.347s
pipeline load max GPU memory allocated: 5.966GiB
pipeline load max GPU memory reserved: 6.025GiB
avg inference time: 4.115s
avg inference time per output: 1.372s
avg inference max GPU memory allocated: 17.233GiB
avg inference max GPU memory reserved: 21.129GiB
```

**With Safety Checker (GPU, safety_check set):**

```bash
pipeline load time: 2.404s
pipeline load max GPU memory allocated: 5.959GiB
pipeline load max GPU memory reserved: 6.045GiB
avg inference time: 4.220s
avg inference time per output: 1.407s
avg inference max GPU memory allocated: 17.225GiB
avg inference max GPU memory reserved: 21.148GiB
```

**With Safety Checker (CPU, safety_check not set):**

```bash
pipeline load time: 2.297s
pipeline load max GPU memory allocated: 4.833GiB
pipeline load max GPU memory reserved: 4.900GiB
avg inference time: 4.200s
avg inference time per output: 1.400s
avg inference max GPU memory allocated: 16.100GiB
avg inference max GPU memory reserved: 20.004GiB
```

**With Safety Checker (CPU, safety_check set):**

```bash
pipeline load time: 2.259s
pipeline load max GPU memory allocated: 4.833GiB
pipeline load max GPU memory reserved: 4.900GiB
avg inference time: 4.860s
avg inference time per output: 1.620s
avg inference max GPU memory allocated: 16.100GiB
avg inference max GPU memory reserved: 20.004GiB
```

#### Conclusion

From the results, we can conclude that the new feature:

- Introduction of the new feature increases the pipeline startup time by approximately 0.5 seconds.
- Enabling `safety_check` increases inference time by about 20ms on the GPU.
- Enabling `safety_check` on the CPU increases inference time significantly, by about 200ms.
- Increases VRAM usage with 1102 MiB when done on the GPU.