@rickstaa 2024-05-06
@yondonfu, @eliteprox, @ad-astra-video, this feature was introduced following a request from CJ on Discord, addressing concerns about NSFW images being generated via their apps. I’d appreciate your insights on integrating this safety check into our pipelines. Additionally, I'm considering whether to include an input check in the `text-to-video` pipeline or if it should remain the responsibility of the dApp. What are your thoughts on this matter?