@dob 2024-05-07
Without reviewing the code, my instinct would be that it would be an option at the pipeline (job type) level, so that users can decide whether they want the filter on or off. My first thought is that the filter would be on by default unless an app decides to override that by passing in the argument to disable the check.
________________________________
From: Rick Staa ***@***.***>
Sent: Monday, May 6, 2024 8:18:39 AM
To: livepeer/ai-worker ***@***.***>
Cc: Doug Petkanics ***@***.***>; Mention ***@***.***>
Subject: Re: [livepeer/ai-worker] feat(text-to-image): integrate NSFW safety checker (PR #76)


@dob<https://github.com/dob>, @yondonfu<https://github.com/yondonfu>, @eliteprox<https://github.com/eliteprox>, @ad-astra-video<https://github.com/ad-astra-video>, this feature was introduced following a request from CJ on Discord, addressing concerns about NSFW images being generated via their apps. I’d appreciate your insights on integrating this safety check into our pipelines. Additionally, I'm considering whether to include an input check in the text-to-video pipeline or if it should remain the responsibility of the dApp. What are your thoughts on this matter?

—
Reply to this email directly, view it on GitHub<https://github.com/livepeer/ai-worker/pull/76#issuecomment-2095749232>, or unsubscribe<https://github.com/notifications/unsubscribe-auth/AAACRFJDVGCUL3UY3DC5QODZA5YJ7AVCNFSM6AAAAABHH5K6ISVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZDAOJVG42DSMRTGI>.
You are receiving this because you were mentioned.Message ID: ***@***.***>